修复浮点型隐式转换成整型丢失小数导致ROOT_NODE与CANVAS_NODE尺寸不一致问题

+++ adapter/ohos/entrance/ace_container.cpp
@@ -934,7 +934,7 @@ void AceContainer::SetView(AceView* view, double density, int32_t width, int32_t
 }
 
 void AceContainer::SetViewNew(
-    AceView* view, double density, int32_t width, int32_t height, sptr<OHOS::Rosen::Window> rsWindow)
+    AceView* view, double density, double width, double height, sptr<OHOS::Rosen::Window> rsWindow)
 {
 #ifdef ENABLE_ROSEN_BACKEND
     CHECK_NULL_VOID(view);
@@ -1383,8 +1383,8 @@ void AceContainer::AddLibPath(int32_t instanceId, const std::vector<std::string>
     assetManagerImpl->SetLibPath("default", libPath);
 }
 
-void AceContainer::AttachView(std::shared_ptr<Window> window, AceView* view, double density, int32_t width,
-    int32_t height, uint32_t windowId, UIEnvCallback callback)
+void AceContainer::AttachView(std::shared_ptr<Window> window, AceView* view, double density, double width,
+    double height, uint32_t windowId, UIEnvCallback callback)
 {
     aceView_ = view;
     auto instanceId = aceView_->GetInstanceId();
